Refactor package matching code
