+++ cmd/gb/internal/match/import.go
@@ -1,6 +1,7 @@
-package main
+package match
 
 import (
+	"fmt"
 	"path"
 	"path/filepath"
 	"strings"
@@ -9,14 +10,13 @@ import (
 )
 
 type Context interface {
-	Projectdir() string
 	AllPackages(string) ([]string, error)
 }
 
 // importPathsNoDotExpansion returns the import paths to use for the given
 // command line, but it does no ... expansion.
-func importPathsNoDotExpansion(ctx Context, cwd string, args []string) []string {
-	srcdir, _ := filepath.Rel(filepath.Join(ctx.Projectdir(), "src"), cwd)
+func importPathsNoDotExpansion(srcdir string, cwd string, args []string) []string {
+	srcdir, _ = filepath.Rel(srcdir, cwd)
 	debug.Debugf("%s %s", cwd, srcdir)
 	if srcdir == ".." {
 		srcdir = "."
@@ -32,30 +32,21 @@ func importPathsNoDotExpansion(ctx Context, cwd string, args []string) []string 
 		if filepath.Separator == '\\' {
 			a = strings.Replace(a, `\`, `/`, -1)
 		}
-
-		if a == "all" || a == "std" {
-			pkgs, err := ctx.AllPackages(a)
-			if err != nil {
-				fatalf("could not load all packages: %v", err)
-			}
-			out = append(out, pkgs...)
-			continue
-		}
 		a = path.Join(srcdir, path.Clean(a))
 		out = append(out, a)
 	}
 	return out
 }
 
-// importPaths returns the import paths to use for the given command line.
-func importPaths(ctx Context, cwd string, args []string) []string {
-	args = importPathsNoDotExpansion(ctx, cwd, args)
+// ImportPaths returns the import paths to use for the given command line.
+func ImportPaths(srcdir string, ctx Context, cwd string, args []string) []string {
+	args = importPathsNoDotExpansion(srcdir, cwd, args)
 	var out []string
 	for _, a := range args {
 		if strings.Contains(a, "...") {
 			pkgs, err := ctx.AllPackages(a)
 			if err != nil {
-				fatalf("could not load all packages: %v", err)
+				fmt.Println("could not load all packages: %v", err)
 			}
 			out = append(out, pkgs...)
 			continue
